[SK-3] change Profile model, fix image deleting from database

+++ backend/users/signals.py
@@ -3,6 +3,7 @@ import os
 from django.conf import settings
 from django.db.models.signals import post_save, pre_save
 from django.dispatch import receiver
+from django.shortcuts import get_object_or_404
 
 from users.models import get_file_path, Profile, User
 
@@ -15,6 +16,10 @@ def create_profile_signal(sender, instance, created, **kwargs):
 
 @receiver(pre_save, sender=Profile)
 def remove_old_image(sender, instance, **kwargs):
-    file_path = get_file_path(instance, instance.file)
-    if os.path.exists(os.path.join(settings.MEDIA_ROOT, file_path)):
-        os.remove(os.path.join(settings.MEDIA_ROOT, file_path))
+
+    current_profile = get_object_or_404(sender, pk=instance.pk)
+
+    if not current_profile.file == instance.file:
+        file_path = get_file_path(instance, instance.file)
+        if os.path.exists(os.path.join(settings.MEDIA_ROOT, file_path)):
+            os.remove(os.path.join(settings.MEDIA_ROOT, file_path))
